+++ tests/qunit/ext.wikispeech.test.util.js
@@ -4,19 +4,21 @@
  * If the div doesn't exist, a new one is added.
  *
  * @param {string} html The HTML added to the div element.
+ * @return {jQuery} The content.
  */
 function setContentHtml( html ) {
 	const contentSelector =
 		mw.config.get( 'wgWikispeechContentSelector' );
+	let $content;
 	if ( $( '#qunit-fixture ' + contentSelector ).length ) {
-		$( '#qunit-fixture ' + contentSelector ).html( html );
+		$content = $( '#qunit-fixture ' + contentSelector );
+		$content.html( html );
 	} else {
-		$( '#qunit-fixture' ).append(
-			$( '<div>' )
-				// Remove the leading "#".
-				.attr( 'id', contentSelector.slice( 1 ) )
-				.html( html )
-		);
+		$content = $( '<div>' )
+			// Remove the leading "#".
+			.attr( 'id', contentSelector.slice( 1 ) );
+		$( '#qunit-fixture' ).append( $content.html( html ) );
 	}
+	return $content;
 }
 module.exports = { setContentHtml };